refactor(Codecov): Move enums to `Config` namespace; rename `Dto` namespace to `Result`; add `CoverageInput`
