+++ tests/Codecov/Unit/Dto/CoverageResultTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Tests\Codecov\Unit\Dto;
 
 use Testo\Assert;
-use Testo\Codecov\Dto\CoverageResult;
-use Testo\Codecov\Dto\LineStatus;
+use Testo\Codecov\Result\CoverageResult;
+use Testo\Codecov\Result\LineStatus;
 use Testo\Test;
 
 #[Test]
