+++ tests/Codecov/Unit/Dto/FunctionCoverageTest.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Tests\Codecov\Unit\Dto;
 
 use Testo\Assert;
-use Testo\Codecov\Dto\BranchCoverage;
-use Testo\Codecov\Dto\FunctionCoverage;
-use Testo\Codecov\Dto\PathCoverage;
+use Testo\Codecov\Result\BranchCoverage;
+use Testo\Codecov\Result\FunctionCoverage;
+use Testo\Codecov\Result\PathCoverage;
 use Testo\Test;
 
 #[Test]
